refactor: Move load_settings function to new settings.py module
